Modernize test and fixture framework
