+++ tests/fixtures/tile_details_response.json
@@ -1,0 +1,107 @@
+{
+  "version": 1,
+  "revision": 1,
+  "timestamp": "2018-06-19T23:04:39.097Z",
+  "timestamp_ms": 1529449479097,
+  "result_code": 0,
+  "result": {
+    "86368462-86fa-4a74-b19e-77f176f0095d": {
+      "thumbnailImage": "https://local-tile-pub.s3.amazonaws.com/..",
+      "tileState": {
+        "ringStateCode": 0,
+        "connectionStateCode": 0,
+        "uuid": "86368462-86fa-4a74-b19e-77f176f0095d",
+        "tile_uuid": "86368462-86fa-4a74-b19e-77f176f0095d",
+        "client_uuid": "2cc56adc-b96a-4293-9b94-eda716e0aa17",
+        "timestamp": 1512615215149,
+        "advertised_rssi": 1.4e-45,
+        "client_rssi": 1.4e-45,
+        "battery_level": 1.4e-45,
+        "latitude": 21.9083423,
+        "longitude": -72.4982138,
+        "altitude": 1821.129812,
+        "h_accuracy": 5,
+        "v_accuracy": 3,
+        "speed": 1.4e-45,
+        "course": 1.4e-45,
+        "authentication": null,
+        "owned": true,
+        "has_authentication": null,
+        "lost_timestamp": -1,
+        "connection_client_uuid": "2cc56adc-b96a-4293-9b94-eda716e0aa17",
+        "connection_event_timestamp": 1512615234268,
+        "last_owner_update": 1512615215149,
+        "connection_state": "READY",
+        "ring_state": "STOPPED",
+        "is_lost": false,
+        "voip_state": "OFFLINE"
+      },
+      "entityName": "TILE",
+      "tile_uuid": "19264d2dffdbca32",
+      "firmware_version": "01.12.14.0",
+      "owner_user_uuid": "2ea56f4d-6576-4b4e-af11-3410cc65e373",
+      "name": "Wallet",
+      "category": null,
+      "image_url": "https://local-tile-pub.s3.amazonaws.com/...",
+      "visible": true,
+      "is_dead": false,
+      "hw_version": "02.09",
+      "product": "DUTCH1",
+      "archetype": "WALLET",
+      "configuration": {
+        "fw10_advertising_interval": null
+      },
+      "last_tile_state": {
+        "ringStateCode": 0,
+        "connectionStateCode": 0,
+        "uuid": "19264d2dffdbca32",
+        "tile_uuid": "19264d2dffdbca32",
+        "client_uuid": "a01bf97a-c89a-40e2-9534-29976010fb03",
+        "timestamp": 1512615215149,
+        "advertised_rssi": 1.4e-45,
+        "client_rssi": 1.4e-45,
+        "battery_level": 1.4e-45,
+        "latitude": 39.797571,
+        "longitude": -104.887826,
+        "altitude": 1588.002773,
+        "h_accuracy": 5,
+        "v_accuracy": 3,
+        "speed": 1.4e-45,
+        "course": 1.4e-45,
+        "authentication": null,
+        "owned": true,
+        "has_authentication": null,
+        "lost_timestamp": -1,
+        "connection_client_uuid": "2cc56adc-b96a-4293-9b94-eda716e0aa17",
+        "connection_event_timestamp": 1512615234268,
+        "last_owner_update": 1512615215149,
+        "connection_state": "DISCONNECTED",
+        "ring_state": "STOPPED",
+        "is_lost": false,
+        "voip_state": "OFFLINE"
+      },
+      "firmware": {
+        "expected_firmware_version": "",
+        "expected_firmware_imagename": "",
+        "expected_firmware_urlprefix": "",
+        "expected_firmware_publish_date": 0,
+        "expected_ppm": null,
+        "expected_advertising_interval": null,
+        "security_level": 1,
+        "expiry_timestamp": 1529471079097,
+        "expected_tdt_cmd_config": null
+      },
+      "auth_key": "aliuUAS7da980asdHJASDQ==",
+      "renewal_status": "LEVEL1",
+      "metadata": {},
+      "auto_retile": false,
+      "status": "ACTIVATED",
+      "tile_type": "TILE",
+      "registration_timestamp": 1482711833983,
+      "is_lost": false,
+      "auth_timestamp": 1512287015405,
+      "activation_timestamp": 1482711835011,
+      "last_modified_timestamp": 1514353410254
+    }
+  }
+}
