+++ tests/fixtures/tile_list_response.json
@@ -1,0 +1,18 @@
+{
+  "version": 1,
+  "revision": 1,
+  "timestamp": "2018-06-19T23:04:32.442Z",
+  "timestamp_ms": 1529449472442,
+  "result_code": 0,
+  "result": [
+    {
+      "tileType": "TILE",
+      "user_uuid": "fd0c10a5-d0f7-4619-9bce-5b2cb7a6754b",
+      "tile_uuid": "86368462-86fa-4a74-b19e-77f176f0095d",
+      "other_user_uuid": "",
+      "other_user_email": "user@email.com",
+      "mode": "OWNER",
+      "last_modified_timestamp": 1482711833985
+    }
+  ]
+}
